Map Refresh Timer Widget
It refreshes the map layers periodically. You can set the preferred map layers and the refresh timer interval. When the timer in up left corner is green, the refresh timer is running and the map layers are being refreshed. By clicking on it, the refresh timer turns off and red. To start the refresh timer, click on it again.